+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.1
+
+- Web support with universal_io
+- Lowers dependency constraint for image
+
+## 1.1.0
+
+- Network and file system image support
+- Algorithm tweaks and bug fixes
+
 ## 1.0.0
 
 - Initial version